Arp modual init
Made a ARP modual able to generate arp noise.

One of its main noise is to do a ARP scan.

Later on in the behavior we will use the __send_arp_request to send randomly request to known host in the network. ( and then sending them other stuff ). Just to add thing to do.