temporary fix for a long term problem
I do not know a solution yet to use telnet in the same shell I generate in my ssh connection.

Maybe in the future I will have one but for now the user gonna connect through SSH to the SMTP server and use telnet to send the message by there.

It is weird.
I dislike it.
I do not know yet how to solve it.