here is the error with windows
Environnement tested on :
1 X windows 10 (with OpenSSH)
1 X windows server 2022 (with OpenSSH and FTP, telnet)
1X Kali Linux ( with OpenSSH, Postfix )

The problem seem to be when a Windows host try to connect via telnet. We do not gain access to that specific shell with our paramiko channel.

added a requirements.txt to help

There is also a need to have Ollama with a AI. But we can remove that functionality for testing.