found couple bug added TODO to correct them
most of them now is mostly timeout from the shell connection ( this shouldnt be to bad to correct :ambulance: ).
Let's hope once this is corrected most of everything work fine and we can start making fun stuff with it.

Last thing to add will be time insted of amount of turn. this will make it easier to work with