WARNING ALOT OF OBJECT MODIFICATIOn
I modified the object to ensure that the object can reconect to the server ( unsure why I didnt do this before ).

Also added smtp_server in the json so no hard coded thing.

This (I hope) should work for saving the shell connection.
I can not run it now because I am on the wrong host.